Replace naive memory management with free-lists.
Cursory performance analysis was showing that the intended API
usage was wasting a lot of time repeatedly mallocing and freeing
small bits of memory for storing each encoded item.

Added a free list for storing previously "freed" encoded items so
that we can re-use already malloced memory instead of freeing it,
just to end up allocating it again shortly afterwards.

Also added a free list for encoded "results", but this is a bit
more awkward as the API for releasing a used result is not
thread-safe -- don't try to release a result back to an encoder
if the result has been transferred to another thread!

The wandder_encode_finish() API has changed as a result of the
performance enhancements, so old code may need to be updated :(

Added new API function wandder_release_encoded_result -- this
can be used to either return a result back to its parent encoder
(and therefore go onto the free list for possible re-use) or to
just destroy the result structure if the encoder given is NULL.