Use inet_ntop to "stringify" IP address fields in ETSI records.
Normally, binary IP address structures are encoded as octet
strings which do not print nicely. However, we can instead flag
these fields with a custom interpretas value
'WANDDER_TAG_BINARY_IP' which will instead cause the field to be
parsed using inet_ntop() -- resulting in a readable IP address.

Also, IA5 and UTF8 strings are now supported by the decoder. For
now, we simply treat them as character strings and trust that the
internal handling of unicode characters is capable.