Try to optimise generalised TS to timeval conversions.
Specifically, we want to avoid continuously calling localtime
and mktime, as these functions are expensive.

Fortunately, we can assume that the result of mktime is not
going to change if the original timestring has not changed, so
we can cache the tv_sec value and just reuse it until the
timestring changes.

This gets us down to a max of 1 mktime() call per second rather
than 1 mktime() call per received ETSI packet.