Added new API function for releasing batches of encoded results.
The provided results must be a linked list and we require pointers
to the head and tail of the list. This will allow users to
release batches of results in a single function call, rather than
one call per result.

Added a mutex for the encoder's freelist to protect against
concurrency issues. The mutex is often tried and if
we cannot grab it then we fallback immediately to using
malloc/free instead of the freelist -- this prevents us from
blocking on the mutex too often. Batch releases do require
a lock on the mutex though.