Fix bad memcpys when encoding reusing old job structures.
If we had an IPPACKET to encode and had assigned that job to a
structure we'd gotten off our freelist, we would incorrectly
decide to memcpy the value contents from the valspace allocated
to the recycled structure.

However, the recycled structure's valspace may be smaller than
the length of the IP packet, so we'd copy random data past the
end of valspace buffer.