Fix leak of wandder items when wandder_decode_skip() is used
The "tree" of cached items was not being built correctly once
a skip had taken place, due to an incorrect "parent" value for
the current decoder post-skip. This meant that a bunch of items
that had been malloc'd were not linked in the cache tree, and
the blob that they belonged to would never be available for reuse.

Also added a couple of simple checks when freeing the cached
item tree to make sure we don't hang if a loop is somehow created.
Note that a loop **should never** happen, but it's probably better
to leak some items than hang entirely.