Fix decoding bug for HI1 Operations messages
There are two different types of "CommunicationIdentifier" in
the ETSI standards -- we were using the wrong one for decoding
HI1 operations messages, because we only had decoding support
for the communicationIdentifier seen in PS headers.

Now I've added support for the other communicationIdentifier and
updated the descend rules for the HI1 notification sequence to
use that instead of the incorrect one.