Add sanity checks to payload decryption
The sanity checks allow us to try and infer whether the
provided decryption key was correct by checking if the
first couple of fields in the decrypted payload container look
sensible.

If not, we can instead warn the user and treat the container
as still being encrypted so that any callers don't end up trying
to interpret the incorrectly decrypted data.